s1_orbits tests
