+++ test/test_GUNW.py
@@ -22,7 +22,6 @@ from RAiDER.aria.prepFromGUNW import (
     check_hrrr_dataset_availablity_for_s1_azimuth_time_interpolation,
     check_weather_model_availability,
 )
-from RAiDER.s1_orbits import _ensure_orbit_credential, ESA_CDSE_HOST
 from RAiDER.cli.raider import calcDelaysGUNW
 from RAiDER.models.customExceptions import *
 
@@ -372,7 +371,7 @@ def test_provenance_metadata_for_tropo_group(weather_model_name: str,
     out_path = shutil.copy(gunw_azimuth_test, tmp_path / out)
 
     if interp_method == 'azimuth_time_grid':
-        mocker.patch('hyp3lib.get_orb.downloadSentinelOrbitFile',
+        mocker.patch('RAiDER.s1_azimuth_timing.downloadSentinelOrbitFile',
                      # Hyp3 Lib returns 2 values
                      side_effect=[
                                   # For azimuth time
@@ -538,76 +537,3 @@ def test_value_error_for_file_inputs_when_no_data_available(mocker):
     with pytest.raises(NoWeatherModelData):
         calcDelaysGUNW(iargs)
     RAiDER.aria.prepFromGUNW.main.assert_not_called()
-
-
-def test__ensure_orbit_credential(monkeypatch):
-    class EmptyNetrc():
-        def __init__(self, netrc_file):
-            self.netrc_file = netrc_file
-            self.hosts = {}
-        def __str__(self):
-            return str(self.hosts)
-
-    # No .netrc, no ESA CDSE env variables
-    with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', EmptyNetrc, raising=False)
-        mp.delenv('ESA_USERNAME', raising=False)
-        mp.delenv('ESA_PASSWORD', raising=False)
-        with pytest.raises(ValueError):
-            _ensure_orbit_credential()
-
-    # No .netrc, set ESA CDSE env variables
-    with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', EmptyNetrc, raising=False)
-        mp.setenv('ESA_USERNAME', 'foo')
-        mp.setenv('ESA_PASSWORD', 'bar')
-        mp.setattr(Path, 'write_text', lambda self, write_text: write_text)
-        written_credentials = _ensure_orbit_credential()
-        assert written_credentials == str({ESA_CDSE_HOST: ('foo', None, 'bar')})
-
-    class NoCDSENetrc():
-        def __init__(self, netrc_file):
-            self.netrc_file = netrc_file
-            self.hosts = {'fizz.buzz.org': ('foo', None, 'bar')}
-        def __str__(self):
-            return str(self.hosts)
-
-    # No CDSE in .netrc, no ESA CDSE env variables
-    with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', NoCDSENetrc, raising=False)
-        mp.delenv('ESA_USERNAME', raising=False)
-        mp.delenv('ESA_PASSWORD', raising=False)
-        with pytest.raises(ValueError):
-            _ensure_orbit_credential()
-
-    # No CDSE in .netrc, set ESA CDSE env variables
-    with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', NoCDSENetrc, raising=False)
-        mp.setenv('ESA_USERNAME', 'foo')
-        mp.setenv('ESA_PASSWORD', 'bar')
-        mp.setattr(Path, 'write_text', lambda self, write_text: write_text)
-        written_credentials = _ensure_orbit_credential()
-        assert written_credentials == str({'fizz.buzz.org': ('foo', None, 'bar'), ESA_CDSE_HOST: ('foo', None, 'bar')})
-
-    class CDSENetrc():
-        def __init__(self, netrc_file):
-            self.netrc_file = netrc_file
-            self.hosts = {ESA_CDSE_HOST: ('foo', None, 'bar')}
-        def __str__(self):
-            return str(self.hosts)
-
-    # cdse in .netrc, no ESA CDSE env variables
-    with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', CDSENetrc, raising=False)
-        mp.delenv('ESA_USERNAME', raising=False)
-        mp.delenv('ESA_PASSWORD', raising=False)
-        written_credentials = _ensure_orbit_credential()
-        assert written_credentials is None
-
-    # cdse in .netrc, set ESA CDSE env variables
-    with monkeypatch.context() as mp:
-        mp.setattr(netrc, 'netrc', CDSENetrc, raising=False)
-        mp.setenv('ESA_USERNAME', 'foo')
-        mp.setenv('ESA_PASSWORD', 'bar')
-        written_credentials = _ensure_orbit_credential()
-        assert written_credentials is None
