+++ tools/RAiDER/aria/prepFromGUNW.py
@@ -23,7 +23,7 @@ from RAiDER.logger import logger
 from RAiDER.models import credentials
 from RAiDER.models.hrrr import HRRR_CONUS_COVERAGE_POLYGON, AK_GEO, check_hrrr_dataset_availability
 from RAiDER.s1_azimuth_timing import get_times_for_azimuth_interpolation
-from RAiDER.s1_orbits import _ensure_orbit_credential
+from RAiDER.s1_orbits import _ensure_orbit_credentials
 
 ## cube spacing in degrees for each model
 DCT_POSTING = {'HRRR': 0.05, 'HRES': 0.10, 'GMAO': 0.10, 'ERA5': 0.10, 'ERA5T': 0.10}
@@ -276,7 +276,7 @@ class GUNW:
         sat = slc.split('_')[0]
         dt  = datetime.strptime(f'{self.dates[0]}T{self.mid_time}', '%Y%m%dT%H:%M:%S')
 
-        _ensure_orbit_credential()
+        _ensure_orbit_credentials()
         path_orb = eof.download.download_eofs([dt], [sat], save_dir=orbit_dir)
 
         return [str(o) for o in path_orb]
